Fix address locator for 1.14a+
