+++ SlashGaming-Diablo-II-API/src/cxx/backend/game_address_table/game_address_locator/game_offset_locator.hpp
@@ -43,20 +43,39 @@
  *  work.
  */
 
-#ifndef SGMAPI_CXX_BACKEND_GAME_ADDRESS_LOCATOR_GAME_ADDRESS_LOCATOR_INTERFACE_HPP_
-#define SGMAPI_CXX_BACKEND_GAME_ADDRESS_LOCATOR_GAME_ADDRESS_LOCATOR_INTERFACE_HPP_
+#ifndef SGMAPI_CXX_BACKEND_GAME_ADDRESS_TABLE_GAME_ADDRESS_LOCATOR_GAME_OFFSET_LOCATOR_HPP_
+#define SGMAPI_CXX_BACKEND_GAME_ADDRESS_TABLE_GAME_ADDRESS_LOCATOR_GAME_OFFSET_LOCATOR_HPP_
 
-#include "../../../../include/cxx/game_address.hpp"
+#include <cstddef>
+#include <filesystem>
+
+#include "game_address_locator.hpp"
 
 namespace mapi {
 
-class IGameAddressLocator {
+class GameOffsetLocator : public IGameAddressLocator {
  public:
-  virtual ~IGameAddressLocator() = default;
+  GameOffsetLocator() = delete;
 
-  virtual GameAddress LocateGameAddress() = 0;
+  GameOffsetLocator(
+      DefaultLibrary library_id,
+      std::ptrdiff_t offset
+  );
+
+  GameOffsetLocator(
+      std::filesystem::path library_path,
+      std::ptrdiff_t offset
+  );
+
+  ~GameOffsetLocator() override;
+
+  GameAddress LocateGameAddress() override;
+
+ private:
+  std::filesystem::path library_path_;
+  std::ptrdiff_t offset_;
 };
 
 } // namespace mapi
 
-#endif // SGMAPI_CXX_BACKEND_GAME_ADDRESS_LOCATOR_GAME_ADDRESS_LOCATOR_INTERFACE_HPP_
+#endif // SGMAPI_CXX_BACKEND_GAME_ADDRESS_TABLE_GAME_ADDRESS_LOCATOR_GAME_OFFSET_LOCATOR_HPP_
